Use mocks to simulate configuring Stripe-related auth
We can use mocks to partially test this part of the code, since it can't be properly tested locally. It will almost certainly need to be revisited once the rest of the system is ready.
In particular, some of the code that was copied from internal-turnkey might not make sense here.

Tests still actually hit Stripe's servers: environment variables are used to store secret keys etc.