Refactor: fix exports/imports for stubbing in tests
If 'standalone' functions are imported into tests with 'import * as thing from <path>', Sinon can stub them: sinon.stub(thing, 'methodName'). This commit moves functions that need to be stubbed back to that arrangement.

Also moved the utility functions for inserting transactions & transactions step to the database into a separate file to reduce the size of executeTransactionPlan.