Add giftbitUserId and identifiers for other payment rails to Stripe charge metadata
Other payment rails: adding identifiers for these provides an easy way for merchants to go from reviewing a Stripe charge in Stripe (ie for fraud), to choosing what to do with any other payment sources that were used in the same transaction (refunding/freezing/etc). The same information could be found by looking up the transaction in Lightrail but storing the information here lets the merchant skip that step.

giftbitUserId: this is not used for anything yet but in the future will make it easier for us to respond to Stripe webhooks to take action on an existing transaction -- ie, when a merchant refunds a transaction and marks it as fraudulent in Stripe, we can rollback some/all of the other steps of a split tender payment, freeze associated Values, etc.

Tests note: test transaction IDs updated to keep Stripe idempotency happy -- charge metadata has changed