Fix tests: skip 'mixed checkout' if stripe env vars missing
Also includes fix to unsetStubsForStripeTests(): the 'if' statements are a check to see if the function in question has been stubbed.

Context: The check for the required env vars happens in the 'before' hook. If the result of the check is that the tests should be skipped, the rest of the code in the 'before' is not run. This means that the stubs are not set, which (without this new check) causes the 'unsetStubsForStripeTests' call in the 'after' hook to fail.