Update stubbed Stripe response with right 'application' property
This seems to be the real 'application' id associated with our test connected/merchant account, so incoming webhook events for example will have this id instead of the previous one. May or may not matter for any current tests but it could be a pain to deal with later (mocked vs live discrepency).