Don't rely on Sentry for error handling
Log error if Stripe calls fail because the credentials can't be fetched (this allows the stack to be deployed in a new environment where the appropriate permissions haven't been set yet). If subsequent Stripe calls fail, for example because we're attempting to enable events with invalid names, allow the error to be thrown and interrupt the deploy process.