Clearer error handling: cases where Stripe should/not retry
If we can't find the Lightrail userId based on the event or if we can't find a corresponding Transaction, it likely won't help for stripe to resend the event.
If some other error occurs, eg we fail to freeze all the Values, Stripe should resend the event so we can try again.