Improve handling of generic values
If the only Values charged were generic, the previous db query in 'freezeValues()' would return 0 and cause the function to throw an error. Added logic to check for this case and log & exit cleanly instead, and also add logging for case where some but not all charged values were generic.