Values report limiting/pagination: 'errorOnOverLimit' replaced by 'suppressLimitError'
Matches change made for Transactions reports in previous commit:
Updated default behaviour: instead of switching on whether a 'limit' param was provided in the request, default to erroring if the query result contains more rows than the limit (or default limit if no limit param provided). Query param 'suppressLimitError=true' overrides this and allows the user to get 'limit' number of rows back regardless of whether there are more.