Rename DbInstanceParameterGroup to trigger new resource creation
During testing for enabling performance_schema and Performance Insights for prod, a number of related parameters were automatically set in this parameter group in dev (parameters which have default values in newly created parameter groups, but didn't when the stack was initially created). Resetting/unsetting the performance_schema related parameters broke dev deploys because the extra parameters were expecting performance_schema etc to still be defined. This change causes the existing parameter group to be replaced with a new one with the new defaults defined.