Refactor Value-fetching query
This project: https://app.getflow.com/organizations/395653/teams/402816/lists/2349808

This query was the most expensive part of handling checkouts.

Replacing the OR WHEREs with UNIONs lets us make better use of our indexes. For most queries that means going from thousands or hundreds of thousands of rows to tens of rows or less.

Removing the GROUP BY Value.id changes the functionality to something we want (a shared generic value can be used by multiple contacts in the same transaction) and more importantly gets rid of a filesort: with the other query improvements this shouldn't matter as much but it's still better to avoid.