Test re-org & naive coverage of edge cases
Tests as documentation for current behaviour for cases where a value is not attached to a contact in the root transaction, then gets attached before the capture/void/reverse happens. Naive interpretation of whether a contact should be associated with a transaction or not - reflects state of the world at the time of the transaction.