Add 'exclude' handling to query for fetching values by contactId
Previously these filters were applied to the full query for fetching all possible values. They were removed so that if a value that could have been charged for the transaction was attached to a contact, that value wold be returned so that we could tag the transaction with the contactId. When fetching values by contactId, we obviously already have the contactId so there's no reason not to keep this part of the query as efficient as possible by applying the filters here. (Any values supplied by code or id still need to be returned even if they cannot be transacted against, but those lookups are already efficient.)