Make migration safer
Overview:
- Adds migrated flag to ContactValues are marked as migrated rather than deleted. This will give us more opportunity to undo changes should anything go horribly wrong.
- Adds a test for generic codes that had been attached two ways (shared + attachGenericAsNewValue: true).
- Nuked package-lock and node modules to fix npm version issues (for whatever reason, when I was npm installing that project I was getting a non-secure url for readable-stream).

Note: Tana had remembered that a long time ago it was possible to update whether a Value was a genericCode or not through the changeCode endpoint. As such, I queried the database to see if there were any ContactValues attached to a Value that's no longer a generic code. There were not any for any environments which means I didn't need to add code to handle this circumstance.