(Merge fix) Fetch all possible values to get all contactIds for transaction in getValuesWithAllContactIds()
This was a previous change that was undone by merging staging: filtering out non-transactable values in the db query makes sense where we can. In order to make sure we track all contactIds that could be involved in the transaction, we can only apply that filter to the part of the db query that looks values up by contactId (in that case we clearly already know what the contactId is). Values referenced by code or valueId must all be returned from the initial db query so that we can tell whether or not they're attached to a contact.

Also fixed a typo in a test and linted.