Simplify getting contactIds for checkout tx tags
Uses existing function getLightrailValuesForTransactionPlanSteps() to fetch values, but using 'nonTransactableHandling: include' so that all values that could possibly be used in the transaction are returned. This allows us to get contactIds from attached values that won't actually be charged *without* making another round trip to the db. Including these contactIds is part of the COS for associating transactions with contacts.